Append streetNumber to "street" when needed
Cart can be configured to have an additional StreetNumber input field. In case its not empty it should be appended to the street input itself. PayOne itself will throw an error in their submit form when no streetnumber is inside the street string, also its an important data for the correct payment information.